chore(ai): dedupe workout analysis prompt builders (CW-392)
buildWorkoutAnalysisData, buildWorkoutAnalysisPrompt, buildAnalysisFactsPromptBlock,
buildAnalysisGuardrailInstructions, getWorkoutTypeGuidance, getAnalysisSectionsGuidance
and normalizeRunningCadence were defined twice -- in trigger/analyze-workout.ts and in
server/utils/services/workoutAnalysisService.ts -- and had already drifted apart. They
now live once in server/utils/services/workout-analysis-prompt.ts and are consumed by
both entry points.

The trigger copy was chosen as canonical: it carries every substantive fix (CW-196 unit
preferences, CW-376 lap-type re-derivation, signal applicability + guardrails), while the
service copy was a stripped-down fork from the Redis-dispatcher commit that was never
resynced.

Also removes the dead V1 buildWorkoutAnalysisFacts computation and the analysisFacts
parameter from buildWorkoutAnalysisPrompt: neither prompt body ever read it, so it was a
wasted pass over the workout streams on every analysis run.
