+++ tests/unit/trigger/analyze-workout-prompt.test.ts
@@ -135,7 +135,6 @@ describe('buildWorkoutAnalysisPrompt', () => {
       },
       null,
       undefined,
-      undefined,
       analysisFactsV2
     )
